+++ crates/store/src/lib.rs
@@ -1,6 +1,4 @@
-use std::time::Duration;
-
-pub mod accounts;
+mod accounts;
 mod blocks;
 mod db;
 mod errors;
@@ -9,13 +7,10 @@ mod inner_forest;
 mod server;
 pub mod state;
 
-pub(crate) use accounts::{AccountTreeWithHistory, HistoricalError};
+pub use accounts::{AccountTreeWithHistory, HistoricalError, InMemoryAccountTree};
 pub use genesis::GenesisState;
 pub use server::{DataDirectory, Store};
 
 // CONSTANTS
 // =================================================================================================
 const COMPONENT: &str = "miden-store";
-
-/// How often to run the database maintenance routine.
-const DATABASE_MAINTENANCE_INTERVAL: Duration = Duration::from_secs(24 * 60 * 60);